refactor: Move mutations items into a dedicated submodule
Doing some cleanup of lib.rs in `pint-abi-gen` before adding array and
keys constructors.
