+++ pint-abi-gen/src/lib.rs
@@ -16,13 +16,14 @@
 //! including the encoding of keys, values and mutations from higher-level types.
 
 use pint_abi_types::{ContractABI, KeyedTypeABI, KeyedVarABI, PredicateABI, TupleField, TypeABI};
-use pint_abi_visit::{tree_from_keyed_vars, KeyedVarTree, Nesting, NodeIx};
+use pint_abi_visit::Nesting;
 use proc_macro::TokenStream;
 use proc_macro2::Span;
 use quote::ToTokens;
 use syn::parse_macro_input;
 
 mod map;
+mod mutation;
 mod tuple;
 mod vars;
 
@@ -179,34 +180,6 @@ fn items_from_predicates(predicates: &[PredicateABI]) -> Vec<syn::Item> {
     items
 }
 
-/// The builder struct for mutations.
-fn mutations_struct() -> syn::ItemStruct {
-    syn::parse_quote! {
-        /// A builder for a set of mutations.
-        ///
-        /// Can be constructed via the [`mutations`] function.
-        #[derive(Debug, Default)]
-        pub struct Mutations {
-            /// The set of mutations being built.
-            set: Vec<pint_abi::types::essential::solution::Mutation>,
-            /// The stack of key elements that need to be merged with the
-            /// `&[Nesting]` derived by the `KeyedTypeABI` traversal.
-            ///
-            /// For example, when a map's `entry` method is called, the provided
-            /// key is pushed to this stack. Upon completion of the `entry`
-            /// method, the key is popped.
-            key_elems: Vec<pint_abi::key::Elem>,
-        }
-    }
-}
-
-fn mutation_method_doc_str(name: &str) -> String {
-    format!(
-        "Add a mutation for the `{name}` field into the set.\n\n\
-        Relaces any existing mutation for the given key.",
-    )
-}
-
 /// Given a keyed var `Nesting`, create a Rust expression that results in its value.
 fn nesting_expr(nesting: &[Nesting]) -> syn::ExprArray {
     let elems = nesting
@@ -244,42 +217,12 @@ fn construct_key_expr() -> syn::Expr {
     }
 }
 
-/// A `Mutations` builder method for a single-key mutation.
-fn mutation_method_for_single_key(
-    name: &str,
-    arg_ty: &SingleKeyTy,
-    nesting: &[Nesting],
-) -> syn::ImplItemFn {
-    let method_ident = syn::Ident::new(name, Span::call_site());
-    let abi_key_doc_str = nesting_key_doc_str(nesting);
-    let doc_str = format!(
-        "{}\n\nKey: `{abi_key_doc_str}`",
-        mutation_method_doc_str(name)
-    );
-    let arg_ty = arg_ty.syn_ty();
-    let nesting_expr: syn::ExprArray = nesting_expr(nesting);
-    let construct_key_expr: syn::Expr = construct_key_expr();
-    syn::parse_quote! {
-        #[doc = #doc_str]
-        pub fn #method_ident(mut self, val: #arg_ty) -> Self {
-            use pint_abi::types::essential::{Key, Value, solution::Mutation};
-            let nesting = #nesting_expr;
-            let key: Key = #construct_key_expr;
-            let value: Value = pint_abi::encode(&val);
-            self.set.retain(|m: &Mutation| &m.key != &key);
-            let mutation = Mutation { key, value };
-            self.set.push(mutation);
-            self
-        }
-    }
-}
-
 /// Convert the given type `&[Nesting]` into a string to use for a generated
 /// builder struct name.
 ///
 /// E.g. `[Var { ix: 1 }, MapEntry { key_ty: i64 }, TupleField { ix: 3, .. }]`
 /// becomes `Var1_MapEntry_Tuple3` so that it may be appended to a builder type
-/// name, e.g. `Tuple_Var1_MapEntry_Tuple3`.
+/// name, e.g. `Tuple_1_MapEntry_3`.
 fn nesting_ty_str<'a>(nesting: impl IntoIterator<Item = &'a Nesting>) -> String {
     fn elem_str(nesting: &Nesting) -> String {
         match nesting {
@@ -338,169 +281,12 @@ fn abi_key_from_keyed_type(ty: &KeyedTypeABI) -> &Vec<Option<usize>> {
     }
 }
 
-/// A `DerefMut<Target = Mutations>` impl for the tuple struct.
-///
-/// This allows for sharing the `mutation_method_from_keyed_var` for method
-/// generation between the `Mutations` impl and the `Tuple_*` impls.
-fn mutation_impl_deref(struct_name: &str) -> Vec<syn::ItemImpl> {
-    let struct_ident = syn::Ident::new(struct_name, Span::call_site());
-    let deref_impl = syn::parse_quote! {
-        impl<'a> core::ops::Deref for #struct_ident<'a> {
-            type Target = Mutations;
-            fn deref(&self) -> &Self::Target {
-                &*self.mutations
-            }
-        }
-    };
-    let deref_mut_impl = syn::parse_quote! {
-        impl<'a> core::ops::DerefMut for #struct_ident<'a> {
-            fn deref_mut(&mut self) -> &mut Self::Target {
-                &mut *self.mutations
-            }
-        }
-    };
-    vec![deref_impl, deref_mut_impl]
-}
-
-/// Recursively traverse the given keyed var and create a builder struct for each tuple.
-fn mutations_builders_from_node(tree: &KeyedVarTree, n: NodeIx) -> Vec<syn::Item> {
-    let mut items = vec![];
-    match tree[n].ty {
-        KeyedTypeABI::Map {
-            ty_from,
-            ty_to,
-            key: _,
-        } => {
-            items.extend(map::builder_items(tree, n, ty_from, ty_to));
-        }
-        KeyedTypeABI::Tuple(_fields) => {
-            items.extend(tuple::builder_items(tree, n));
-        }
-        _ => (),
-    }
-    items
-}
-
-/// Recursively traverse the given keyed vars and create a builder structs and
-/// impls for each tuple.
-fn mutations_builders_from_keyed_vars(vars: &[KeyedVarABI]) -> Vec<syn::Item> {
-    let mut items = vec![];
-    let tree = pint_abi_visit::tree_from_keyed_vars(vars);
-    tree.dfs(|n| {
-        items.extend(mutations_builders_from_node(&tree, n));
-    });
-    items
-}
-
-/// A `Mutations` builder method for a map field.
-fn mutation_method_for_map(name: &str, map_nesting: &[Nesting]) -> syn::ImplItemFn {
-    let struct_name = map::mutations_struct_name(map_nesting);
-    let method_ident = syn::Ident::new(name, Span::call_site());
-    let struct_ident = syn::Ident::new(&struct_name, Span::call_site());
-    let doc_str = mutation_method_doc_str(name);
-    syn::parse_quote! {
-        #[doc = #doc_str]
-        pub fn #method_ident(mut self, f: impl FnOnce(#struct_ident) -> #struct_ident) -> Self {
-            f(#struct_ident { mutations: &mut self });
-            self
-        }
-    }
-}
-
-/// A `Mutations` builder method for a tuple field.
-fn mutation_method_for_tuple(name: &str, nesting: &[Nesting]) -> syn::ImplItemFn {
-    let struct_name = tuple::mutations_struct_name(nesting);
-    let method_ident = syn::Ident::new(name, Span::call_site());
-    let struct_ident = syn::Ident::new(&struct_name, Span::call_site());
-    let doc_str = mutation_method_doc_str(name);
-    syn::parse_quote! {
-        #[doc = #doc_str]
-        pub fn #method_ident(mut self, f: impl FnOnce(#struct_ident) -> #struct_ident) -> Self {
-            f(#struct_ident { mutations: &mut self });
-            self
-        }
-    }
-}
-
-/// A `Mutations` builder method for the given keyed var.
-fn mutation_method_from_node(tree: &KeyedVarTree, n: NodeIx, name: &str) -> syn::ImplItemFn {
-    let nesting = tree.nesting(n);
-    let arg_ty = match &tree[n].ty {
-        KeyedTypeABI::Bool(_key) => SingleKeyTy::Bool,
-        KeyedTypeABI::Int(_key) => SingleKeyTy::Int,
-        KeyedTypeABI::Real(_key) => SingleKeyTy::Real,
-        KeyedTypeABI::String(_key) => SingleKeyTy::String,
-        KeyedTypeABI::B256(_key) => SingleKeyTy::B256,
-        KeyedTypeABI::Array { ty: _, size: _ } => todo!(),
-        // Tuple types take a closure.
-        KeyedTypeABI::Tuple(_) => {
-            return mutation_method_for_tuple(name, &nesting);
-        }
-        // Map types take a closure.
-        KeyedTypeABI::Map { .. } => {
-            return mutation_method_for_map(name, &nesting);
-        }
-    };
-    // A mutation builder method for a single mutation.
-    mutation_method_for_single_key(name, &arg_ty, &nesting)
-}
-
-/// All builder methods for the `Mutations builder type.
-fn mutations_methods_from_keyed_vars(vars: &[KeyedVarABI]) -> Vec<syn::ImplItemFn> {
-    let tree = tree_from_keyed_vars(vars);
-    tree.roots()
-        .iter()
-        .map(|&n| {
-            let name = field_name_from_var_name(tree[n].name.unwrap());
-            mutation_method_from_node(&tree, n, &name)
-        })
-        .collect()
-}
-
-/// The implementation for the `Mutations` builder type.
-fn impl_mutations_from_keyed_vars(vars: &[KeyedVarABI]) -> syn::ItemImpl {
-    let methods = mutations_methods_from_keyed_vars(vars);
-    syn::parse_quote! {
-        impl Mutations {
-            #(
-                #methods
-            )*
-        }
-    }
-}
-
-/// Shorthand constructor for the `Mutations` builder.
-fn mutations_fn() -> syn::ItemFn {
-    syn::parse_quote! {
-        /// Begin building a set of [`Mutations`].
-        pub fn mutations() -> Mutations {
-            Mutations::default()
-        }
-    }
-}
-
-/// The `From<Mutations>` implementation for `Vec<Mutation>`.
-fn impl_from_mutations() -> syn::ItemImpl {
-    syn::parse_quote! {
-        impl From<Mutations> for Vec<pint_abi::types::essential::solution::Mutation> {
-            fn from(m: Mutations) -> Self {
-                m.set
-            }
-        }
-    }
-}
-
 /// The `mutations` and `keys` items for the given keyed vars.
 ///
 /// This is used for both `storage` and `pub_vars` mod generation.
 fn items_from_keyed_vars(vars: &[KeyedVarABI]) -> Vec<syn::Item> {
-    let mut items: Vec<syn::Item> = vec![
-        mutations_struct().into(),
-        mutations_fn().into(),
-        impl_from_mutations().into(),
-        impl_mutations_from_keyed_vars(vars).into(),
-    ];
-    items.extend(mutations_builders_from_keyed_vars(vars));
+    let mut items = vec![];
+    items.extend(mutation::items(vars));
     items
 }
